@@ -72,31 +73,31 @@
     "date-fns": "^3.6.0",
     "dotenv": "^16.5.0",
     "drizzle-orm": "^0.44.5",
-    "drizzle-zod": "^0.8.2",
+    "drizzle-zod": "^0.8.3",
     "embla-carousel-react": "^8.6.0",
     "express": "^4.21.2",
-    "express-rate-limit": "^7.5.0",
+    "express-rate-limit": "^8.1.0",
     "express-session": "^1.18.1",
     "express-validator": "^7.2.1",
-    "framer-motion": "^12.15.0",
+    "framer-motion": "^12.23.22",
     "helmet": "^8.1.0",
     "input-otp": "^1.4.2",
-    "joi": "^17.13.3",
+    "joi": "^18.0.1",
     "ldapts": "^8.0.9",
     "lucide-react": "^0.511.0",
     "memorystore": "^1.6.7",
     "multer": "^2.0.0",
     "nanoid": "^5.0.8",
-    "newrelic": "^13.3.2",
+    "newrelic": "^13.3.3",
     "next-themes": "^0.4.6",
     "nodemailer": "^7.0.3",
     "passport": "^0.7.0",
     "passport-local": "^1.0.0",
-    "pg": "^8.15.6",
+    "pg": "^8.16.3",
     "puppeteer": "^24.15.0",
-    "react": "^18.3.1",
+    "react": "^19.1.1",
     "react-day-picker": "^8.10.1",
-    "react-dom": "^18.3.1",
+    "react-dom": "^19.1.1",
     "react-hook-form": "^7.56.1",
     "react-icons": "^5.5.0",
     "react-resizable-panels": "^2.1.9",
@@ -129,19 +130,19 @@
     "@types/passport": "^1.0.16",
     "@types/passport-local": "^1.0.38",
     "@types/pg": "^8.11.10",
-    "@types/react": "^18.3.23",
-    "@types/react-dom": "^18.3.7",
+    "@types/react": "^19.1.13",
+    "@types/react-dom": "^19.1.9",
     "@types/ws": "^8.5.13",
     "@vitejs/plugin-react": "^4.5.0",
     "autoprefixer": "^10.4.20",
     "babel-plugin-react-remove-properties": "0.3.0",
     "drizzle-kit": "^0.31.1",
-    "esbuild": "^0.25.5",
+    "esbuild": "^0.25.10",
     "postcss": "^8.5.4",
-    "tailwindcss": "^3.4.17",
+    "tailwindcss": "^4.1.13",
     "tsx": "^4.19.1",
-    "typescript": "^5.6.3",
-    "vite": "^6.3.5"
+    "typescript": "^5.9.2",
+    "vite": "^6.3.6"
   },
   "optionalDependencies": {
     "bufferutil": "^4.0.8"
