@@ -96,4 +96,4 @@ const config = {
   }
 };
 
-export = config;
+export default config;
